LayoutTests/media/session/content-interruptions-expected.txt

'Content' start of interruption events should indefinitely pause 'Content' media sessions.


Waiting for Content media to begin playing.
Media began playing.
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
Sending 'Content' start-of-interruption notification.
RUN(internals.sendMediaSessionStartOfInterruptionNotification("content"))
The media element was paused.
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
END OF TEST
